moved factory.go to discovery/loaders.go
