+++ internal/controller/targetsource_controller.go
@@ -31,7 +31,6 @@ import (
 	gnmicv1alpha1 "github.com/gnmic/operator/api/v1alpha1"
 	"github.com/gnmic/operator/internal/controller/discovery"
 	"github.com/gnmic/operator/internal/controller/discovery/core"
-	"github.com/gnmic/operator/internal/controller/discovery/loaders"
 	"github.com/go-logr/logr"
 )
 
@@ -210,7 +209,7 @@ func (r *TargetSourceReconciler) startDiscoveryPipeline(key types.NamespacedName
 
 	// Create loader instance
 	if loaderConfigured {
-		loader, err := loaders.NewLoader(
+		loader, err := discovery.NewLoader(
 			key,
 			targetSource.Spec,
 			core.LoaderConfig{ChunkSize: r.ChunkSize},
